sniffer

+++ IntWarsSharp/Logic/Packets/Packets.cs
@@ -1394,18 +1394,18 @@ namespace IntWarsSharp.Logic.Packets
             else
                 buffer.Write((byte)0x40); // unk -- seems to be flags related to things like critical strike (0x49)
 
-            buffer.Write(0x40);
-            buffer.Write(0x01);
-            buffer.Write(0x7B);
-            buffer.Write(0xEF);
-            buffer.Write(0xEF);
-            buffer.Write(0x01);
-            buffer.Write(0x2E);
-            buffer.Write(0x55);
-            buffer.Write(0x55);
-            buffer.Write(0x35);
-            buffer.Write(0x94);
-            buffer.Write(0xD3);
+            buffer.Write((byte)0x40);
+            buffer.Write((byte)0x01);
+            buffer.Write((byte)0x7B);
+            buffer.Write((byte)0xEF);
+            buffer.Write((byte)0xEF);
+            buffer.Write((byte)0x01);
+            buffer.Write((byte)0x2E);
+            buffer.Write((byte)0x55);
+            buffer.Write((byte)0x55);
+            buffer.Write((byte)0x35);
+            buffer.Write((byte)0x94);
+            buffer.Write((byte)0xD3);
         }
     }
